운동 세션 관련 API 및 UI 개선: 세션 데이터 수정 기능(무게,횟수) 추가, 사용자 인증 로직 통합 미들웨어 추가가, 코드 리팩토링 및 가독성 향상
