+++ app/api/user/exerciseSession/[sessionId]/route.ts
@@ -1,9 +1,6 @@
 import ExerciseSession from "@/models/ExerciseSession"
-import connect from "@/utils/db"
 import { NextRequest, NextResponse } from "next/server"
-import { getServerSession } from "next-auth"
-import User from "@/models/User"
-import { getToken } from "next-auth/jwt"
+import { requireUser } from "@/lib/check-auth"
 
 /**
  *  운동 세션 id로 조회
@@ -13,19 +10,10 @@ import { getToken } from "next-auth/jwt"
 
 export const GET = async (req: NextRequest, { params }: { params: Promise<{ sessionId: string }> }) => {
     const sessionId = (await params).sessionId; // 요청에서 사용자 ID 가져오기
-    const getSession = await getServerSession();
-    const token = await getToken({ req, secret: process.env.NEXTAUTH_SECRET });
-
-    if (!getSession || !token) {
-        // 로그인 안되어있으면 로그인 페이지로 이동
-        return NextResponse.redirect(`${process.env.NEXTAUTH_URL}/login`);
-    }
     try {
-        await connect();
-        const user = await User.findOne({ email: getSession.user.email, provider: token.provider });
-        if (!user) {
-            return NextResponse.json({ message: 'User not found' }, { status: 404 });
-        }
+        const { user, error, status } = await requireUser(req);
+        if (!user) return NextResponse.json({ message: error }, { status });
+
         const exerciseSession = await ExerciseSession.findOne({ _id: sessionId });
         return NextResponse.json(exerciseSession, { status: 200 });
     } catch (err: any) {
@@ -34,22 +22,13 @@ export const GET = async (req: NextRequest, { params }: { params: Promise<{ sess
 }
 
 
-export const PATCH = async (req: NextRequest, { params }: { params: Promise<{ sessionId: string }> }) => {
-    const sessionId = (await params).sessionId; // 요청에서 사용자 ID 가져오기
-
-    const getSession = await getServerSession();
-    const token = await getToken({ req, secret: process.env.NEXTAUTH_SECRET });
-
-    if (!getSession || !token) {
-        // 로그인 안되어있으면 로그인 페이지로 이동
-        return NextResponse.redirect(`${process.env.NEXTAUTH_URL}/login`);
-    }
+export const PATCH = async (req: NextRequest) => {
+    const { sessionId } = await req.json(); // 요청에서 사용자 ID 가져오기
     try {
-        await connect();
-        const user = await User.findOne({ email: getSession.user.email, provider: token.provider });
-        if (!user) {
-            return NextResponse.json({ message: 'User not found' }, { status: 404 });
-        }
+        const { user, error, status } = await requireUser(req);
+        if (!user) return NextResponse.json({ message: error }, { status });
+
+
         const checkSession = await ExerciseSession.findOne({ _id: sessionId });
         if (!checkSession) {
             return NextResponse.json({ message: '존재하지 않는 세션입니다.' }, { status: 404 });
