+++ server/mutations.ts
@@ -1,6 +1,6 @@
 
 import { toast } from "@/hooks/use-toast";
-import { Exercise, ExercisePlan, ExerciseSession, ExerciseSessionActionPayload, ExercisesessionData, PostExerciseSession } from "@/utils/util";
+import { Exercise, ExercisePlan, ExerciseSession, ExerciseSessionActionPayload, ExercisesessionData, ExerciseSessionSetPayload, PostExerciseSession } from "@/utils/util";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 
 /**
@@ -313,13 +313,13 @@ export const useDetailDeletePlan = () => {
 
 /** 
  * 운동 세션 생성 Mutation
+ * @description 운동 세션을 생성하는 Mutation입니다. 운동 계획에 따라 세션을 생성합니다.
+ * @param exerciseSession - 운동 세션의 정보를 담고 있는 객체입니다.
 */
-
 export const useCreateExerciseSession = () => {
     const queryClient = useQueryClient();
     return useMutation({
         mutationFn: async (exerciseSession: ExerciseSession) => {
-
             const response = await fetch(`${process.env.NEXTAUTH_URL}/api/user/exerciseSession`, {
                 method: 'POST',
                 headers: { 'Content-Type': 'application/json', },
@@ -331,31 +331,26 @@ export const useCreateExerciseSession = () => {
             }
             return response.json();
         },
-        onSuccess: () => {
-            console.log('onSuccess');
+        onSuccess: async (data) => {
+            toast({ variant: "default2", title: "운동을 시작합니다!" });
+            await queryClient.invalidateQueries({ queryKey: ["exerciseSession", data.newExerciseSession._id] }) // 데이터 갱신 후 자동으로 UI 업데이트
+            await queryClient.invalidateQueries({ queryKey: ["inProgress"] }) // 데이터 갱신 후 자동으로 UI 업데이트
         },
         onError: (error) => {
-            console.log('onError', error);
+            const message = error instanceof Error ? error.message : String(error);
+            console.error("onError", error);
+            toast({ variant: "destructive", title: message });
         },
-        onSettled: async (data, error) => { // 성공, 실패 상관없이 마지막에 호출 variables
-            console.log('onSettled');
-            if (error) {
-                toast({ variant: 'destructive', title: `${error}` });
-                console.log('error', error);
-            } else {
-                await queryClient.invalidateQueries({ queryKey: ["exerciseSession", data.newExerciseSession._id] }) // 데이터 갱신 후 자동으로 UI 업데이트
-                await queryClient.invalidateQueries({ queryKey: ["inProgress"] }) // 데이터 갱신 후 자동으로 UI 업데이트
-            }
-        }
     })
 };
 
 
 /**
  * 운동 세션 상태 변경 Mutation
+ * @description 운동 세션의 상태를 변경하는 Mutation입니다. 운동 시작, 종료, 세트 완료 등의 작업을 수행합니다.
+ * @param data - 운동 세션의 ID와 운동 ID, 상태 변경 작업('start' || 'done')을 포함하는 객체입니다.
  * @returns 
  */
-
 export const useActionExerciseSession = () => {
     const queryClient = useQueryClient();
     return useMutation({
@@ -435,11 +430,11 @@ export const useAExerciseSession = () => {
 export const useEditExerciseSession = () => {
     const queryClient = useQueryClient();
     return useMutation({
-        mutationFn: async (data: any) => {
-            const response = await fetch(`${process.env.NEXTAUTH_URL}/api/user/exerciseSession/${data.sessionId}/${data.exerciseId}`, {
+        mutationFn: async (payload: ExerciseSessionSetPayload) => {
+            const response = await fetch(`${process.env.NEXTAUTH_URL}/api/user/exerciseSession/${payload.sessionId}/${payload.exerciseId}`, {
                 method: 'PATCH',
                 headers: { 'Content-Type': 'application/json', },
-                body: JSON.stringify(data)
+                body: JSON.stringify(payload)
             });
             if (!response.ok) {
                 const errorData = await response.json();
@@ -447,18 +442,17 @@ export const useEditExerciseSession = () => {
             }
             return response.json();
         },
-        onSettled: async (data, error) => { // 성공, 실패 상관없이 마지막에 호출 variables
-            console.log('onSettled');
-            if (error) {
-                toast({ variant: 'destructive', title: `${error}` });
-                console.log('error', error);
-            } else {
-                console.log('data', data);
-                toast({ variant: 'default2', title: `${data.message}` });
-                await queryClient.invalidateQueries({ queryKey: ["exerciseSession", data.updatedSession._id] }) // 데이터 갱신 후 자동으로 UI 업데이트
-                await queryClient.invalidateQueries({ queryKey: ["inProgress"] }) // 데이터 갱신 후 자동으로 UI 업데이트
-            }
-        }
+        onSuccess: async (data) => {
+            toast({ variant: 'default2', title: data.message });
+            await Promise.all([
+                queryClient.invalidateQueries({ queryKey: ['exerciseSession', data.updatedSession._id] }),
+                queryClient.invalidateQueries({ queryKey: ['inProgress'] }),
+            ]);
+        },
+        onError: (error: any) => {
+            console.error('Error updating session:', error);
+            toast({ variant: 'destructive', title: error.message || '업데이트 중 에러 발생' });
+        },
     })
 };
