+++ utils/db.ts
@@ -5,7 +5,6 @@ const connect = async () => {
 
   try {
     await mongoose.connect(process.env.MONGO_URI!, {});
-    console.log("Mongo Connection successfully established.");
   } catch (error) {
     throw new Error("Error connecting to Mongoose");
   }
